MCP_INFO_MGR/requirements_semantic_search.txt
# Semantic Search Dependencies for MCP Tool Discovery
# Install with: pip install -r MCP_INFO_MGR/requirements_semantic_search.txt

# Vector search
faiss-cpu==1.9.0  # For CPU-only systems
# faiss-gpu==1.9.0  # Uncomment for GPU support (requires CUDA)

# Embedding model
sentence-transformers==3.3.1

# Deep learning backend
torch>=2.0.0

# Numerical computing
numpy>=1.24.0

# Note: These dependencies are for the semantic search indexing and search features.
# They are separate from the main Orchestrator dependencies.